+++ examples/simplest.js
@@ -1,18 +1,20 @@
-const { loadModel, transcript, freeModel } = require('../voskjs')
+const { logLevel, loadModel, transcript, freeModel } = require('../voskjs')
 
 async function main() {
 
   const modelDirectory = '../models/vosk-model-en-us-aspire-0.2'
   const audioFile = '../audio/2830-3980-0043.wav'
 
-  console.log(`model directory         : ${modelDirectory}`)
-  console.log(`speech file name        : ${audioFile}`)
-  console.log()
+  console.log(`model directory      : ${modelDirectory}`)
+  console.log(`speech file name     : ${audioFile}`)
+
+  // set the vosk log level to silence 
+  logLevel(-1) 
 
   // load in memory a Vosk directory model
   const { model, latency } = await loadModel(modelDirectory)
 
-  console.log(`init model latency   : ${latency}ms`)
+  console.log(`load model latency   : ${latency}ms`)
 
   // speech recognition of an audio file
   try {